cloud_functions/nuke_cache/requirements.txt
google-cloud-secret-manager==2.8.0
redis==3.5.3
